requirements.txt
pillow
numpy
protobuf
# just for the colormaps
matplotlib
# (if you want to use 3d visualization utils)
# open3d
# (if you wwant to use the .csv-reading function)
# pandas
# pyyaml
# (if you wwant to use the clustering function)
# sklearn